requirements.txt
flask
scikit-learn
langdetect
deep-translator
pytesseract
Pillow
moviepy
docx2txt
PyPDF2
openai-whisper
joblib